Pull out ReleasesApiConsumer to its own file
